Updates the level commands (#133)
* Adds xp and level management commands

Adds the 3 new commands to manage the xp of a user and the level of a user.

* Fixes argument notation in level commands

Corrects the argument notation in the level commands documentation.
Specifically, changes from square brackets to angle brackets for
required arguments in `addxp` and `setlevel` commands to clarify
that these arguments are mandatory.

* Updates the merged level commands

Consolidates color, accent, and opacity settings into the level config command and modifies the level reset command to include server resets.
Also renames level shh/dm commands to toggle.